Milestones 1-2, large file support, custom scrollbar, design journal
Complete plain-text editor with custom rendering engine:
- PieceTable with IBuffer abstraction (string, lazy file, procedural)
- All offsets promoted int→long for large file support
- Windowed layout via ILogicalScrollable for 1M+ line documents
- Two-zone DualZoneScrollBar replacing Avalonia ScrollViewer
- Recent files menu with dev-mode procedural sample documents
- File open/save with FileLoader/FileSaver
- 69 tests passing (48 Core + 21 Rendering)

Design journal captures architectural direction: WYSIWYG block tree
model, variable height strategy, Fenwick tree height index, wrap modes.

+++ src/DevMentalMd.Core/Document/History/DeleteEdit.cs
@@ -3,7 +3,7 @@ namespace DevMentalMd.Core.Documents.History;
 /// <summary>Records a deletion so it can be applied or reverted.</summary>
 /// <param name="Ofs">Logical document offset where the deletion started.</param>
 /// <param name="DeletedText">The text that was removed (needed to reconstruct on revert).</param>
-public sealed record DeleteEdit(int Ofs, string DeletedText) : IDocumentEdit {
+public sealed record DeleteEdit(long Ofs, string DeletedText) : IDocumentEdit {
     public void Apply(PieceTable table) => table.Delete(Ofs, DeletedText.Length);
     public void Revert(PieceTable table) => table.Insert(Ofs, DeletedText);
 }
